Skip milliseconds in the timestamp since they cause problems

+++ migrations/20170724154717_init.js
@@ -5,7 +5,7 @@ exports.up = function(knex, Promise) {
       table.increments('id').primary();
       table.string('slackId');
       table.string('message');
-      table.float('timestamp');
+      table.integer('timestamp');
     }));
 };
